Update the strainID
