+++ install.sh
@@ -15,7 +15,7 @@ NB=`wc -l < $LDIR/.tmp`
 
 
 g++ -std=c++11 -o $LDIR/src/genSpecies $LDIR/src/genSpecies.cpp -O3
-g++ -std=c++11 -o $LDIR/src/buildDB  $LDIR/src/compress.cpp -O3
+g++ -std=c++11 -o $LDIR/src/buildDB  $LDIR/src/DTB_compress.cpp -O3
 g++ -std=c++11 -o $LDIR/src/uniqueDB  $LDIR/src/DTB_unique.cpp -O3
 g++ -std=c++11 -o $LDIR/src/translate  $LDIR/src/translate.cpp -O3
 g++ -fopenmp -std=c++11 -o $LDIR/src/classify  $LDIR/src/classify.cpp $LDIR/src/seqreader.cpp -O3
